Create neuron and layer packages (#172)
Summary:
Since the number of algorithms under `attr_/core` is increasing moving layer and neuron variants into `layer` and `neuron` packages correspondingly. This will not change the import signature for any of the algorithms.
In addition to that I'm also adding shortcuts for the missing algorithms in `__init__.py`
Pull Request resolved: https://github.com/pytorch/captum/pull/172

Differential Revision: D18535972

Pulled By: NarineK

fbshipit-source-id: bfff632d617f712cb40b9759cc604bd221f0df1f