Extend interpretable embedding with multiple args (#191)
Summary:
This PR extends InterpretableEmbeddingBase to be more generic and support nested embeddings of any types and number of inputs.
I've also added test cases and documentation related to this PR.

In addition to that it looks like in the IMDB tutorial the device was missing and some minor fixes. Adding that as well.
Pull Request resolved: https://github.com/pytorch/captum/pull/191

Differential Revision: D18779108

Pulled By: NarineK

fbshipit-source-id: aa4819a49592ea028c5845defd5b8b5f076eb28c