Feature/permutation feature importance (#185)
Summary:
Implementation of blackbox feature permutation importance algorithm: https://christophm.github.io/interpretable-ml-book/feature-importance.html

Pseudo-code of the above, for a single batch of data:

```
perm_feature_importance(batch):
     feature_importance = {}
     initial_err = f(batch)
     for feature in features:
           batch_copy = permute_feature_across(batch, feature)
           err = f(batch_copy)
           diff = initial_err - err
           feature_importance[feature] = diff

     return feature_importance

permute_feature_across(batch, feature):
      batch_copy = batch.clone()
      batch_copy[:, feature] = batch[random_perm(n), feature]
      return batch_copy
```

The basic idea is to permute each feature in a batch and see how different the output (or error/loss) is from the output/error/loss of the original batch. If the error increases => we know that this feature is important. If the error remains about the same => we know that the feature is not that important since the model was not very reliant on it. Finally, if the error decreases => we know likely know that the feature was negatively impacting the model's performance and hence was detrimental.

error/loss functions must be provided in the `forward_func`. See the test cases for examples.
Pull Request resolved: https://github.com/pytorch/captum/pull/185

Differential Revision: D19148113

Pulled By: miguelmartin75

fbshipit-source-id: bd1868b9dc7a283b01fdd9211cad8edcd7605dc6